[RN] refactor: create separate component SwipeToDelete
