+++ src/components/containers/SwipeToDeleteContainer/index.ts
@@ -1,0 +1,1 @@
+export {SwipeToDeleteContainer} from './SwipeToDeleteContainer';
