+++ src/components/containers/SwipeToDeleteContainer/styles.ts
@@ -1,0 +1,12 @@
+import {StyleSheet} from 'react-native';
+import {COLORS} from 'assets';
+
+export const styles = StyleSheet.create({
+  rightDeleteActionBox: {
+    flex: 1,
+    justifyContent: 'center',
+    alignItems: 'flex-end',
+    backgroundColor: COLORS.persimmon,
+    paddingRight: 20,
+  },
+});
