+++ src/components/containers/index.ts
@@ -1,1 +1,2 @@
 export {FavoriteButtonContainer} from './FavoriteButtonContainer';
+export {SwipeToDeleteContainer} from './SwipeToDeleteContainer';
